fixing memory leak with 1.7
