+++ web/func.go
@@ -6,8 +6,6 @@ import (
 	"net/http"
 	"strconv"
 	"time"
-
-	"github.com/gorilla/context"
 )
 
 // Let's have generic errors for expected conditions. Typically
@@ -103,27 +101,3 @@ func ParseTruthyFalsy(flag interface{}) (result bool, err error) {
 	}
 	return strconv.ParseBool(s)
 }
-
-// Vars is a helper function for accessing route
-// parameters from any server.Router implementation. This is the equivalent
-// of using `mux.Vars(r)` with the Gorilla mux.Router.
-func Vars(r *http.Request) map[string]string {
-	if rv := context.Get(r, varsKey); rv != nil {
-		return rv.(map[string]string)
-	}
-	return nil
-}
-
-// SetRouteVars will set the given value into into the request context
-// with the shared 'vars' storage key.
-func SetRouteVars(r *http.Request, val interface{}) {
-	if val != nil {
-		context.Set(r, varsKey, val)
-	}
-}
-
-type contextKey int
-
-// key to set/retrieve URL params from a
-// Gorilla request context.
-const varsKey contextKey = 2
